Someone didn't like it randomly ending

+++ chiru/cogs/myfun.py
@@ -115,11 +115,14 @@ class MyFun(object):
                         return
 
     @commands.command(pass_context=True)
-    async def loadingbar(self, ctx: Context, length: int = 10, message: str = "Loading", fail: bool = False,
+    async def loadingbar(self, ctx: Context, length: int = 10, message: str = "Loading", fail_at: int = 0,
                          hide: bool = False, after: str = ""):
         """
         Displays a loading bar, actually doesn't load anything.
         """
+        fail = False
+        if fail_at!=0:
+            fail = True
         await self.bot.delete_message(ctx.message)
         if hide:
             title = ""
@@ -134,7 +137,6 @@ class MyFun(object):
                                                                 message if fail else message, "\u2588" * i,
                                                                 "_" * (length - i)))
             await asyncio.sleep(1)
-            if fail and i > length // 3 and randint(1, length // 3) == 1:
                 break
         if not fail:
             message = message[0].lower() + message[1:]
